fix: foodTagIds의 제네릭을 Integer에서 Long으로 변경

+++ BE/src/main/java/com/domain/user/dto/request/EaterSignUpRequest.java
@@ -27,7 +27,7 @@ public record EaterSignUpRequest(
         String nickname,
 
         @Schema(description = "선호 음식", example = "[1, 4, 9]")
-        List<Integer> foodTagIds,
+        List<Long> foodTagIds,
 
         @Size(max = 50, message = "CUSTOM_FOOD_TAG_TOO_LONG")
         @Schema(description = "사용자 정의 선호 음식", example = "마라탕")
